requirements.txt
matplotlib
opencv-python
scikit-image
torchvision
torch
Pillow